+++ WkndRay/Materials/AbstractMaterial.cs
@@ -10,9 +10,9 @@ namespace WkndRay.Materials
     {
         public abstract ScatterResult Scatter(Ray rayIn, HitRecord hitRecord);
 
-        public virtual double ScatteringPdf(Ray rayIn, HitRecord hitRecord, Ray scattered)
+        public virtual float ScatteringPdf(Ray rayIn, HitRecord hitRecord, Ray scattered)
         {
-            return 0.0;
+            return 0.0f;
         }
 
         public virtual ColorVector Emitted(Ray rayIn, HitRecord hitRecord, Point2D uvCoords, PosVector p)
