+++ WkndRay/Materials/IMaterial.cs
@@ -9,7 +9,7 @@ namespace WkndRay.Materials
     public interface IMaterial
     {
         ScatterResult Scatter(Ray rayIn, HitRecord hitRecord);
-        double ScatteringPdf(Ray rayIn, HitRecord hitRecord, Ray scattered);
+        float ScatteringPdf(Ray rayIn, HitRecord hitRecord, Ray scattered);
         ColorVector Emitted(Ray rayIn, HitRecord hitRecord, Point2D uvCoords, PosVector p);
     }
 }
