+++ WkndRay/Scenes/ImageTextureScene.cs
@@ -22,16 +22,16 @@ namespace WkndRay.Scenes
         /// <inheritdoc />
         public Camera GetCamera(int imageWidth, int imageHeight)
         {
-            var lookFrom = new PosVector(13.0, 2.0, 3.0);
+            var lookFrom = new PosVector(13.0f, 2.0f, 3.0f);
             var lookAt = PosVector.Zero;
-            double distToFocus = 10.0;
-            double aperture = 0.0;
+            float distToFocus = 10.0f;
+            float aperture = 0.0f;
             return new Camera(
               lookFrom,
               lookAt,
               PosVector.UnitY,
-              40.0,
-              Convert.ToDouble(imageWidth) / Convert.ToDouble(imageHeight),
+              40.0f,
+              Convert.ToSingle(imageWidth) / Convert.ToSingle(imageHeight),
               aperture,
               distToFocus);
         }
@@ -43,16 +43,16 @@ namespace WkndRay.Scenes
             var list = new HitableList()
       {
         new Sphere(
-          new PosVector(0.0, -1000.0, 0.0),
-          1000.0,
-          new LambertianMaterial(new VectorNoiseTexture(VectorNoiseMode.Soft, 3.0))),
+          new PosVector(0.0f, -1000.0f, 0.0f),
+          1000.0f,
+          new LambertianMaterial(new VectorNoiseTexture(VectorNoiseMode.Soft, 3.0f))),
         new Sphere(
-          new PosVector(0.0, 2.0, 0.0),
-          2.0,
+          new PosVector(0.0f, 2.0f, 0.0f),
+          2.0f,
           new LambertianMaterial(new ImageTexture(globe)))
       };
 
-            return new BvhNode(list, 0.0, 1.0);
+            return new BvhNode(list, 0.0f, 1.0f);
         }
 
         /// <inheritdoc />
@@ -67,8 +67,8 @@ namespace WkndRay.Scenes
             return ray =>
             {
                 var unitDirection = ray.Direction.ToUnitVector();
-                double t = 0.5 * (unitDirection.Y + 1.0);
-                return ((1.0 - t) * ColorVector.One) + (t * new ColorVector(0.5, 0.7, 1.0));
+                float t = 0.5f * (unitDirection.Y + 1.0f);
+                return ((1.0f - t) * ColorVector.One) + (t * new ColorVector(0.5f, 0.7f, 1.0f));
             };
         }
     }
